# servers/assessment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the assessment server of PROMPT.",
        "title": "PROMPT Assessment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/assessment/api",
    "paths": {
...
        "/course_phase/{coursePhaseID}/config/participations": {
            "get": {
                "description": "Get course participations for a course phase from core service.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course_phase_config"
                ],
                "summary": "List participations for course phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/coursePhaseConfigDTO.AssessmentParticipationWithStudent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/evaluation/feedback-items/course-participation/{courseParticipationID}": {
            "get": {
                "description": "List feedback items for a course participation in a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "List feedback items for participant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedbackItemDTO.FeedbackItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/student-assessment/action-item/action/course-participation/{courseParticipationID}": {
            "get": {
                "description": "List action items for a course participation in a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action_items"
                ],
                "summary": "List action items for a student (communication)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/actionItemDTO.ActionItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/student-assessment/action-item/course-participation/{courseParticipationID}": {
            "get": {
                "description": "List action items for a course participation in the course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action_items"
                ],
                "summary": "List action items for student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/actionItemDTO.ActionItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/student-assessment/completed/course-participation/{courseParticipationID}": {
            "get": {
                "description": "Get an assessment completion by course participation ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessment_completions"
                ],
                "summary": "Get assessment completion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/assessmentCompletionDTO.AssessmentCompletion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an assessment completion by course participation ID.",
                "tags": [
                    "assessment_completions"
                ],
                "summary": "Delete assessment completion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/student-assessment/completed/grade/course-participation/{courseParticipationID}": {
            "get": {
                "description": "Get grade suggestion for a course participation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessment_completions"
                ],
                "summary": "Get grade for student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/student-assessment/course-participation/{courseParticipationID}": {
            "get": {
                "description": "List assessments for a course participation in the course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "List assessments for student in phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/assessmentDTO.Assessment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "actionItemDTO.ActionItem": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "assessmentCompletionDTO.AssessmentCompletion": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                },
                "completedAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "gradeSuggestion": {
                    "type": "number"
                }
            }
        },
...
        "assessmentDTO.Assessment": {
            "type": "object",
            "properties": {
                "assessedAt": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "competencyID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "examples": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scoreLevel": {
                    "$ref": "#/definitions/scoreLevelDTO.ScoreLevel"
                }
            }
        },
...
        "assessmentType.AssessmentType": {
            "type": "string",
            "enum": [
                "self",
                "peer",
                "tutor",
                "assessment"
            ],
            "x-enum-varnames": [
                "Self",
                "Peer",
                "Tutor",
                "Assessment"
            ]
        },
...
        "coursePhaseConfigDTO.AssessmentParticipationWithStudent": {
            "type": "object",
            "properties": {
                "courseParticipationID": {
                    "description": "CourseParticipationID links this phase participation to the overall course participation.",
                    "type": "string"
                },
                "coursePhaseID": {
                    "description": "CoursePhaseID is the unique identifier of the course phase the student is participating in.",
                    "type": "string"
                },
                "passStatus": {
                    "description": "PassStatus indicates the student's current status in this course phase.\nCommon values are usually \"passed\", \"failed\", and \"not_assessed\".",
                    "type": "string"
                },
                "prevData": {
                    "description": "PrevData contains metadata from previous phases, passed via course-phase communication",
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.MetaData"
                        }
                    ]
                },
                "restrictedData": {
                    "description": "RestrictedData contains sensitive metadata that should only be accessible to authorized users\nsuch as instructors, admins, or the system itself.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.MetaData"
                        }
                    ]
                },
                "student": {
                    "description": "Student contains the complete student information associated with this participation.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.Student"
                        }
                    ]
                },
                "studentReadableData": {
                    "description": "StudentReadableData contains metadata that can be safely shared with the student,\nsuch as feedback, scores, or progress information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.MetaData"
                        }
                    ]
                },
                "teamID": {
                    "type": "string"
                }
            }
        },
...
        "db.FeedbackType": {
            "type": "string",
            "enum": [
                "positive",
                "negative"
            ],
            "x-enum-varnames": [
                "FeedbackTypePositive",
                "FeedbackTypeNegative"
            ]
        },
...
        "feedbackItemDTO.FeedbackItem": {
            "type": "object",
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "feedbackText": {
                    "type": "string"
                },
                "feedbackType": {
                    "$ref": "#/definitions/db.FeedbackType"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/assessmentType.AssessmentType"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Int4": {
            "type": "object",
            "properties": {
                "int32": {
                    "type": "integer",
                    "format": "int32"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "promptTypes.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "diverse",
                "prefer_not_to_say"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderDiverse",
                "GenderPreferNotToSay"
            ]
        },
        "promptTypes.MetaData": {
            "type": "object",
            "additionalProperties": true
        },
...
        "promptTypes.Student": {
            "type": "object",
            "properties": {
                "currentSemester": {
                    "description": "CurrentSemester indicates which semester the student is currently in within their program.\nThis helps with course eligibility, academic planning, and progress tracking.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pgtype.Int4"
                        }
                    ]
                },
                "email": {
                    "description": "Email is the student's contact email address, must be a valid email format.\nThis is used for notifications, communications, and account recovery.",
                    "type": "string"
                },
                "firstName": {
                    "description": "FirstName stores the first name as provided during initial application or profile updates.",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender represents the student's gender identity for demographic and statistical purposes.\nMust be one of: \"male\", \"female\", \"diverse\", or \"prefer_not_to_say\".",
                    "enum": [
                        "male",
                        "female",
                        "diverse",
                        "prefer_not_to_say"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.Gender"
                        }
                    ]
                },
                "hasUniversityAccount": {
                    "description": "HasUniversityAccount indicates whether the student has been granted access to university systems.\nThis flag helps determine what services and resources the student can access.",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID is the unique identifier for the person using UUID v4 format.\nThe uuid.UUID type ensures type safety and automatic validation during JSON unmarshaling.",
                    "type": "string"
                },
                "lastName": {
                    "description": "LastName stores the last name as provided during initial application or profile updates.",
                    "type": "string"
                },
                "matriculationNumber": {
                    "description": "MatriculationNumber is the student's unique university identification number.\nThis number is assigned by the university and remains constant throughout the student's enrollment.",
                    "type": "string"
                },
                "nationality": {
                    "description": "Nationality represents the student's nationality or citizenship.\nThis information may be used for visa requirements, international student services, or statistics.",
                    "type": "string"
                },
                "studyDegree": {
                    "description": "StudyDegree indicates the type of degree the student is pursuing.\nMust be either \"bachelor\" or \"master\".",
                    "enum": [
                        "bachelor",
                        "master"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/promptTypes.StudyDegree"
                        }
                    ]
                },
                "studyProgram": {
                    "description": "StudyProgram is the specific program or major the student is enrolled in.\nExamples: \"Computer Science\", \"Information Systems\", \"Management and Technology\".",
                    "type": "string"
                },
                "universityLogin": {
                    "description": "UniversityLogin is the student's username for university systems and services.\nThis login is typically used for accessing library, computing resources, and other university platforms.",
                    "type": "string"
                }
            }
        },
        "promptTypes.StudyDegree": {
            "type": "string",
            "enum": [
                "bachelor",
                "master"
            ],
            "x-enum-varnames": [
                "StudyDegreeBachelor",
                "StudyDegreeMaster"
            ]
        },
        "scoreLevelDTO.ScoreLevel": {
            "type": "string",
            "enum": [
                "veryBad",
                "bad",
                "ok",
                "good",
                "veryGood"
            ],
            "x-enum-varnames": [
                "ScoreLevelVeryBad",
                "ScoreLevelBad",
                "ScoreLevelOk",
                "ScoreLevelGood",
                "ScoreLevelVeryGood"
            ]
        },
...
    },
...
}